+++ src/wfdir.c
@@ -625,7 +625,7 @@ DirWndProc(
          if (szItem[0] == '\0')
                 szItem = L"..";
          cchMatch = wcslen(rgchMatch);
-         if (cchMatch > (INT)wcslen(szItem))
+         if (cchMatch > wcslen(szItem))
                 cchMatch = wcslen(szItem);
 
          if (CompareString( LOCALE_USER_DEFAULT, NORM_IGNORECASE, 
@@ -2901,7 +2901,7 @@ UsedAltname:
          goto GDSExit;
 
       if ((!bLFNTest) && ((i + 1) < iMac))
-		 if(p)
+         if(p)
             lstrcat(p, szBlank);
    }
